Separate the represented instance and difficulty responsibility (#613)
Ninth delivery of the refactor push. 82 methods move into 5 bounded submodules
under session/instances: instance, binding, difficulty, map_key and
map_resolution. Every file is under the 1,000-line review budget.

session/mod.rs drops from 41,107 to 39,283 production lines. Across the nine
deliveries the root falls from 75,841 to 39,283 - 48 per cent retired into
navigable owners, and the root impl block from 2,598 methods to under 1,000.

map_key first came out at 1,017 lines. Rather than raise a ceiling, the
methods that resolve the canonical map and manager the Session is attached to
moved into their own map_resolution module, which is a distinct operation from
deriving and comparing map keys.

Methods that mention a map or instance but belong to creature, gameobject,
item, spell, aura, movement, transfer, quest, pet or loot responsibilities
already moved with those families and are not touched again.

Safety established by comparison: the impl surface is 1,139 methods before and
after, an identical multiset, and the ownership baseline reports 3,761
associated items unchanged.

Verified: cargo test -p wow-world --lib 3,822 passed and 0 failed,
session-ownership-check --syntax-only PASS, check_architecture.py check and
self-test PASS, fmt and diff hygiene clean. Ceiling tightened.
